+++ src/components/tools/tune.ts
@@ -1,15 +1,15 @@
 import BaseTool from './base';
 import type { BlockAPI, BlockTune as IBlockTune, BlockTuneConstructable } from '@/types';
 import type { BlockTuneData } from '@/types/block-tunes/block-tune-data';
-import type { BlockTuneFactory } from '@/types/tools/wrappers/block-tune-factory';
-import { ToolType } from '@/types/tools/wrappers/tool-type';
+import type { BlockTuneAdapter } from '@/types/tools/adapters/block-tune-adapter';
+import { ToolType } from '@/types/tools/adapters/tool-type';
 
 /**
  * Stub class for BlockTunes
  *
  * @todo Implement
  */
-export default class BlockTune extends BaseTool<ToolType.Tune, IBlockTune> implements BlockTuneFactory {
+export default class BlockTune extends BaseTool<ToolType.Tune, IBlockTune> implements BlockTuneAdapter {
   /**
    * Tool type — Tune
    */
