+++ types/tools/adapters/tool-type.ts
